requirements.txt

# Step 1: Standard
django >= 2.2.4

# Step 2: Decoupling sensitive infos from GIT repo
python-decouple >= 3.1
dj-database-url >= 0.5.0
mysqlclient >= 1.3.13

django-bootstrap4 >= 0.0.6
#requests >= 2.19.1
docutils >= 0.14
django-filter >= 2.0.0
django-widget-tweaks >= 1.4.2
django-tables2 >= 1.21.2
django-debug-toolbar >= 1.9.1
django-import-export >= 1.1.0
tblib >= 1.3.2
xhtml2pdf >= 0.2.3
django-mdeditor >= 0.1.16

djangorestframework>=3.9.2
sqlparse>=0.2.4
requests>=2.21.0
django-environ>=0.4.5
django-crispy-forms>=1.7.2
simplejson>=3.16.0
urllib3>=1.24.1
Pillow>=5.4.1


# Images are at ${Project}./myvenv/lib/python3.5/site-packages/django/contrib/gis/static/gis/img
